report-service/requirements.txt
Flask>=3.0.0
SQLAlchemy>=2.0.0
Flask-SQLAlchemy>=3.1.1
psycopg2-binary>=2.9.0
WeasyPrint>=60.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
